test: extract out helper function to create two test users
